+++ core/changelog.md
@@ -1,4 +1,5 @@
 - fix: allow setting authorization header through extra headers and in allow list and deny list
+- feat: added image generation support for bedrock
 - feat: added support for gemini google search tool
 - fix: function response part handling in gemini
 - fix: call enrich error in vertex to return raw request and response in bifrost errors